#46 Merge api and spec submodules
