+++ tests/test_api.py
@@ -7,11 +7,32 @@ import pandas as pd
 from pandas.testing import assert_frame_equal
 from typespecs import (
     ITSELF,
+    ItselfType,
     Spec,
+    SpecFrame,
     from_annotated,
     from_annotation,
     from_annotations,
+    is_spec,
+    is_specframe,
 )
+
+
+def test_itself() -> None:
+    assert ItselfType() == ITSELF
+    assert ItselfType() is not ITSELF
+
+
+def test_spec() -> None:
+    spec = Spec(a=1, b=2, c=ITSELF)
+    assert is_spec(spec)
+    assert not is_spec({})
+
+
+def test_specframe() -> None:
+    specs = SpecFrame()
+    assert is_specframe(specs)
+    assert not is_specframe(pd.DataFrame())
 
 
 def test_from_annotated() -> None:
